refactor(settings): move provider accounts into dedicated Agent Accounts pane
Extracts the Claude, Codex, Gemini, and OpenCode Go account/provider sections
out of General and into a new "Agent Accounts" sidebar section so related
controls live together and General stays focused on workspace/editor/updates.
