+++ src/renderer/src/components/settings/GeneralPane.tsx
@@ -2,17 +2,12 @@
    splitting individual settings into separate files would scatter related controls without a
    meaningful abstraction boundary. */
 import { useEffect, useRef, useState } from 'react'
-import type {
-  ClaudeRateLimitAccountsState,
-  CodexRateLimitAccountsState,
-  GlobalSettings
-} from '../../../../shared/types'
-import { Badge } from '../ui/badge'
+import type { GlobalSettings } from '../../../../shared/types'
 import { Button } from '../ui/button'
 import { Input } from '../ui/input'
 import { Label } from '../ui/label'
 import { Separator } from '../ui/separator'
-import { Download, FolderOpen, Loader2, Plus, RefreshCw, Star, Timer, Trash2 } from 'lucide-react'
+import { Download, FolderOpen, Loader2, RefreshCw, Star, Timer } from 'lucide-react'
 import { useAppStore } from '../../store'
 import { CliSection } from './CliSection'
 import { toast } from 'sonner'
@@ -23,13 +18,9 @@ import {
 } from '../../../../shared/constants'
 import { clampNumber } from '@/lib/terminal-theme'
 import {
-  GENERAL_CLAUDE_ACCOUNTS_SEARCH_ENTRIES,
-  GENERAL_CODEX_ACCOUNTS_SEARCH_ENTRIES,
   GENERAL_CACHE_TIMER_SEARCH_ENTRIES,
   GENERAL_CLI_SEARCH_ENTRIES,
   GENERAL_EDITOR_SEARCH_ENTRIES,
-  GENERAL_GEMINI_SEARCH_ENTRIES,
-  GENERAL_OPENCODE_SEARCH_ENTRIES,
   GENERAL_PANE_SEARCH_ENTRIES,
   GENERAL_SUPPORT_SEARCH_ENTRIES,
   GENERAL_UPDATE_SEARCH_ENTRIES,
@@ -38,15 +29,6 @@ import {
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../ui/select'
 import { SearchableSetting } from './SearchableSetting'
 import { matchesSettingsSearch } from './settings-search'
-import { markLiveCodexSessionsForRestart } from '@/lib/codex-session-restart'
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle
-} from '../ui/dialog'
 
 export { GENERAL_PANE_SEARCH_ENTRIES }
 
@@ -55,74 +37,9 @@ type GeneralPaneProps = {
   updateSettings: (updates: Partial<GlobalSettings>) => void
 }
 
-function getCodexAccountLabel(
-  state: CodexRateLimitAccountsState,
-  accountId: string | null | undefined
-): string {
-  if (accountId == null) {
-    return 'System default'
-  }
-  return state.accounts.find((account) => account.id === accountId)?.email ?? 'Codex account'
-}
-
-function getClaudeAccountLabel(
-  state: ClaudeRateLimitAccountsState,
-  accountId: string | null | undefined
-): string {
-  if (accountId == null) {
-    return 'System default'
-  }
-  return state.accounts.find((account) => account.id === accountId)?.email ?? 'Claude account'
-}
-
-function getCodexAccountErrorDescription(error: unknown): string {
-  const message = String((error as Error)?.message ?? error)
-    .replace(/^Error occurred in handler for 'codexAccounts:[^']+':\s*/i, '')
-    .replace(/^Error invoking remote method 'codexAccounts:[^']+':\s*/i, '')
-    .replace(/^Error:\s*/i, '')
-    .trim()
-  const normalizedMessage = message.toLowerCase()
-
-  // Why: Codex account actions cross the Electron IPC boundary, and invoke()
-  // failures often include transport-level wrapper text that is useful in
-  // devtools but noisy in product UI. Normalize the handful of expected auth
-  // failures here so users see actionable sign-in guidance instead of IPC
-  // internals or raw upstream wording.
-  if (normalizedMessage.includes('timed out waiting for codex login to finish')) {
-    return 'Codex sign-in took too long to finish. Please try again.'
-  }
-  if (normalizedMessage.includes('codex sign-in took too long to finish')) {
-    return 'Codex sign-in took too long to finish. Please try again.'
-  }
-  if (
-    normalizedMessage.includes('auth error 502') ||
-    normalizedMessage.includes('gateway') ||
-    normalizedMessage.includes('bad gateway')
-  ) {
-    return 'Codex sign-in is temporarily unavailable. Please try again in a minute.'
-  }
-  if (normalizedMessage.startsWith('codex login failed:')) {
-    const loginMessage = message.slice('Codex login failed:'.length).trim()
-    return loginMessage || 'Codex sign-in failed. Please try again.'
-  }
-
-  return message || 'Codex sign-in failed. Please try again.'
-}
-
-function getClaudeAccountErrorDescription(error: unknown): string {
-  return (
-    String((error as Error)?.message ?? error)
-      .replace(/^Error occurred in handler for 'claudeAccounts:[^']+':\s*/i, '')
-      .replace(/^Error invoking remote method 'claudeAccounts:[^']+':\s*/i, '')
-      .replace(/^Error:\s*/i, '')
-      .trim() || 'Claude sign-in failed. Please try again.'
-  )
-}
-
 export function GeneralPane({ settings, updateSettings }: GeneralPaneProps): React.JSX.Element {
   const searchQuery = useAppStore((s) => s.settingsSearchQuery)
   const updateStatus = useAppStore((s) => s.updateStatus)
-  const fetchSettings = useAppStore((s) => s.fetchSettings)
   // Why: the 'error' variant of UpdateStatus does not carry a `version` field.
   // The main process emits `{ state: 'error' }` for both check failures (no
   // version known yet) and download/install failures (version was known from
@@ -151,22 +68,6 @@ export function GeneralPane({ settings, updateSettings }: GeneralPaneProps): Rea
   const [autoSaveDelayDraft, setAutoSaveDelayDraft] = useState(
     String(settings.editorAutoSaveDelayMs)
   )
-  const [codexAccounts, setCodexAccounts] = useState<CodexRateLimitAccountsState>({
-    accounts: [],
-    activeAccountId: null
-  })
-  const [codexAction, setCodexAction] = useState<
-    'idle' | 'adding' | `reauth:${string}` | `remove:${string}` | `select:${string | 'system'}`
-  >('idle')
-  const [claudeAccounts, setClaudeAccounts] = useState<ClaudeRateLimitAccountsState>({
-    accounts: [],
-    activeAccountId: null
-  })
-  const [claudeAction, setClaudeAction] = useState<
-    'idle' | 'adding' | `reauth:${string}` | `remove:${string}` | `select:${string | 'system'}`
-  >('idle')
-  const [removeAccountId, setRemoveAccountId] = useState<string | null>(null)
-  const [removeClaudeAccountId, setRemoveClaudeAccountId] = useState<string | null>(null)
   // Why: the star state is derived from gh, not from settings, so it does not
   // live in the global settings store. 'hidden' covers the gh-unavailable and
   // already-starred-on-a-previous-session cases so the section drops out for
@@ -221,47 +122,6 @@ export function GeneralPane({ settings, updateSettings }: GeneralPaneProps): Rea
     setAutoSaveDelayDraft(String(settings.editorAutoSaveDelayMs))
   }, [settings.editorAutoSaveDelayMs])
 
-  useEffect(() => {
-    let stale = false
-
-    const loadCodexAccounts = async (): Promise<void> => {
-      try {
-        const nextCodex = await window.api.codexAccounts.list()
-        if (!stale) {
-          setCodexAccounts(nextCodex)
-        }
-      } catch (error) {
-        if (!stale) {
-          toast.error('Could not load Codex accounts.', {
-            description: String((error as Error)?.message ?? error)
-          })
-        }
-      }
-    }
-
-    const loadClaudeAccounts = async (): Promise<void> => {
-      try {
-        const nextClaude = await window.api.claudeAccounts.list()
-        if (!stale) {
-          setClaudeAccounts(nextClaude)
-        }
-      } catch (error) {
-        if (!stale) {
-          toast.error('Could not load Claude accounts.', {
-            description: String((error as Error)?.message ?? error)
-          })
-        }
-      }
-    }
-
-    void loadCodexAccounts()
-    void loadClaudeAccounts()
-
-    return () => {
-      stale = true
-    }
-  }, [])
-
   const handleBrowseWorkspace = async () => {
     const path = await window.api.repos.pickFolder()
     if (path) {
@@ -297,79 +157,6 @@ export function GeneralPane({ settings, updateSettings }: GeneralPaneProps): Rea
     // if the IPC channel itself breaks. Log defensively; the user will notice
     // the app didn't restart and can retry.
     void window.api.updater.quitAndInstall().catch(console.error)
-  }
-
-  const syncCodexAccounts = async (next: CodexRateLimitAccountsState): Promise<void> => {
-    setCodexAccounts(next)
-    await fetchSettings()
-  }
-
-  const syncClaudeAccounts = async (next: ClaudeRateLimitAccountsState): Promise<void> => {
-    setClaudeAccounts(next)
-    await fetchSettings()
-  }
-
-  const formatAccountTimestamp = (timestamp: number): string => {
-    return new Date(timestamp).toLocaleString(undefined, {
-      month: 'short',
-      day: 'numeric',
-      hour: 'numeric',
-      minute: '2-digit'
-    })
-  }
-
-  const runCodexAccountAction = async (
-    action: typeof codexAction,
-    operation: () => Promise<CodexRateLimitAccountsState>
-  ): Promise<void> => {
-    const previousActiveAccountId = codexAccounts.activeAccountId
-    setCodexAction(action)
-    try {
-      const next = await operation()
-      await syncCodexAccounts(next)
-      const shouldPromptRestart =
-        action === 'adding' ||
-        (action.startsWith('select:') && previousActiveAccountId !== next.activeAccountId) ||
-        (action.startsWith('reauth:') &&
-          next.activeAccountId !== null &&
-          action === `reauth:${next.activeAccountId}`) ||
-        (action.startsWith('remove:') && previousActiveAccountId !== next.activeAccountId)
-      if (shouldPromptRestart) {
-        void markLiveCodexSessionsForRestart({
-          previousAccountLabel: getCodexAccountLabel(codexAccounts, previousActiveAccountId),
-          nextAccountLabel: getCodexAccountLabel(next, next.activeAccountId)
-        })
-      }
-    } catch (error) {
-      toast.error('Codex account update failed.', {
-        description: getCodexAccountErrorDescription(error)
-      })
-    } finally {
-      setCodexAction('idle')
-    }
-  }
-
-  const runClaudeAccountAction = async (
-    action: typeof claudeAction,
-    operation: () => Promise<ClaudeRateLimitAccountsState>
-  ): Promise<void> => {
-    const previousActiveAccountId = claudeAccounts.activeAccountId
-    setClaudeAction(action)
-    try {
-      const next = await operation()
-      await syncClaudeAccounts(next)
-      if (previousActiveAccountId !== next.activeAccountId || action === 'adding') {
-        toast.info('Claude account updated.', {
-          description: `${getClaudeAccountLabel(claudeAccounts, previousActiveAccountId)} → ${getClaudeAccountLabel(next, next.activeAccountId)}. Restart live Claude terminals before continuing old sessions.`
-        })
-      }
-    } catch (error) {
-      toast.error('Claude account update failed.', {
-        description: getClaudeAccountErrorDescription(error)
-      })
-    } finally {
-      setClaudeAction('idle')
-    }
   }
 
   const visibleSections = [
@@ -670,477 +457,6 @@ export function GeneralPane({ settings, updateSettings }: GeneralPaneProps): Rea
         )}
       </section>
     ) : null,
-    matchesSettingsSearch(searchQuery, GENERAL_CLAUDE_ACCOUNTS_SEARCH_ENTRIES) ? (
-      <section key="claude-accounts" id="general-claude-accounts" className="space-y-4 scroll-mt-6">
-        <div className="space-y-1">
-          <h3 className="text-sm font-semibold">Claude Accounts</h3>
-          <p className="text-xs text-muted-foreground">
-            Add and switch Claude Code accounts without moving chat sessions to account-specific
-            config directories.
-          </p>
-        </div>
-
-        <SearchableSetting
-          title="Claude Accounts"
-          description="Manage which Claude account Orca materializes into the shared Claude auth files."
-          keywords={['claude', 'account', 'rate limit', 'status bar', 'quota']}
-          className="space-y-3 px-1 py-2"
-        >
-          <div className="flex items-center justify-between gap-3">
-            <div className="space-y-0.5">
-              <Label>Accounts</Label>
-              <p className="text-xs text-muted-foreground">
-                Orca swaps Claude auth only; config and chat history stay in the shared Claude root.
-              </p>
-            </div>
-            <Button
-              variant="outline"
-              size="xs"
-              onClick={() =>
-                void runClaudeAccountAction('adding', () => window.api.claudeAccounts.add())
-              }
-              disabled={claudeAction !== 'idle'}
-              className="gap-1.5"
-            >
-              {claudeAction === 'adding' ? (
-                <Loader2 className="size-3 animate-spin" />
-              ) : (
-                <Plus className="size-3" />
-              )}
-              Add Account
-            </Button>
-          </div>
-
-          <div className="space-y-2">
-            <button
-              type="button"
-              onClick={() =>
-                void runClaudeAccountAction('select:system', () =>
-                  window.api.claudeAccounts.select({ accountId: null })
-                )
-              }
-              disabled={claudeAction !== 'idle'}
-              className={`flex w-full items-center justify-between gap-3 rounded-md border px-3 py-2.5 text-left transition-colors ${
-                claudeAccounts.activeAccountId === null
-                  ? 'border-foreground/20 bg-accent/15'
-                  : 'border-border/70 hover:border-border hover:bg-accent/8'
-              }`}
-            >
-              <div className="flex min-w-0 flex-1 flex-col gap-0.5">
-                <div className="flex min-w-0 items-center gap-2">
-                  <span className="truncate text-sm font-medium">System default</span>
-                  {claudeAccounts.activeAccountId === null ? (
-                    <Badge
-                      variant="outline"
-                      className="h-4 shrink-0 rounded px-1.5 text-[10px] font-medium leading-none text-foreground/80"
-                    >
-                      Active
-                    </Badge>
-                  ) : null}
-                </div>
-                <span className="truncate text-[11px] text-muted-foreground">
-                  Use your current system Claude login.
-                </span>
-              </div>
-            </button>
-            {claudeAccounts.accounts.length === 0 ? (
-              <div className="rounded-md border border-dashed border-border/70 px-3 py-4 text-xs text-muted-foreground">
-                No managed Claude accounts yet. Orca will use your system default Claude login until
-                you add one here.
-              </div>
-            ) : (
-              claudeAccounts.accounts.map((account) => {
-                const isActive = claudeAccounts.activeAccountId === account.id
-                const isReauthing = claudeAction === `reauth:${account.id}`
-                const isBusy = claudeAction !== 'idle'
-
-                return (
-                  <button
-                    key={account.id}
-                    type="button"
-                    onClick={() =>
-                      void runClaudeAccountAction(`select:${account.id}`, () =>
-                        window.api.claudeAccounts.select({ accountId: account.id })
-                      )
-                    }
-                    disabled={isBusy}
-                    className={`flex w-full items-center justify-between gap-3 rounded-md border px-3 py-2.5 text-left transition-colors ${
-                      isActive
-                        ? 'border-foreground/20 bg-accent/15'
-                        : 'border-border/70 hover:border-border hover:bg-accent/8'
-                    }`}
-                  >
-                    <div className="flex w-full items-center justify-between gap-3 max-md:flex-col max-md:items-start">
-                      <div className="flex min-w-0 flex-1 flex-col gap-0.5">
-                        <div className="flex min-w-0 items-center gap-2">
-                          <span className="truncate text-sm font-medium">{account.email}</span>
-                          {isActive ? (
-                            <Badge
-                              variant="outline"
-                              className="h-4 shrink-0 rounded px-1.5 text-[10px] font-medium leading-none text-foreground/80"
-                            >
-                              Active
-                            </Badge>
-                          ) : null}
-                        </div>
-                        <span className="truncate text-[11px] text-muted-foreground">
-                          {account.organizationName
-                            ? `${account.organizationName} · ${formatAccountTimestamp(account.lastAuthenticatedAt)}`
-                            : formatAccountTimestamp(account.lastAuthenticatedAt)}
-                        </span>
-                      </div>
-                      <div className="flex shrink-0 items-center justify-end gap-1 max-md:w-full max-md:flex-wrap">
-                        <Button
-                          variant="ghost"
-                          size="xs"
-                          onClick={(event) => {
-                            event.stopPropagation()
-                            void runClaudeAccountAction(`reauth:${account.id}`, () =>
-                              window.api.claudeAccounts.reauthenticate({ accountId: account.id })
-                            )
-                          }}
-                          disabled={isBusy}
-                          className="h-6 px-2 text-muted-foreground hover:text-foreground"
-                        >
-                          {isReauthing ? (
-                            <Loader2 className="size-3 animate-spin" />
-                          ) : (
-                            <RefreshCw className="size-3" />
-                          )}
-                          Re-authenticate
-                        </Button>
-                        <Button
-                          variant="ghost"
-                          size="xs"
-                          onClick={(event) => {
-                            event.stopPropagation()
-                            setRemoveClaudeAccountId(account.id)
-                          }}
-                          disabled={isBusy}
-                          className="h-6 px-2 text-muted-foreground hover:text-destructive"
-                        >
-                          <Trash2 className="size-3" />
-                          Remove
-                        </Button>
-                      </div>
-                    </div>
-                  </button>
-                )
-              })
-            )}
-          </div>
-        </SearchableSetting>
-      </section>
-    ) : null,
-    matchesSettingsSearch(searchQuery, GENERAL_CODEX_ACCOUNTS_SEARCH_ENTRIES) ? (
-      <section key="codex-accounts" id="general-codex-accounts" className="space-y-4 scroll-mt-6">
-        <div className="space-y-1">
-          <h3 className="text-sm font-semibold">Codex Accounts</h3>
-          <p className="text-xs text-muted-foreground">
-            Add and switch between Codex accounts in Orca.
-          </p>
-          <p className="text-xs text-muted-foreground">
-            Each account keeps its own local sign-in context in Orca. Account auth stays on this
-            device.
-          </p>
-        </div>
-
-        <SearchableSetting
-          title="Codex Accounts"
-          description="Manage which Codex account Orca uses for live rate limit fetching."
-          keywords={['codex', 'account', 'rate limit', 'status bar', 'quota']}
-          className="space-y-3 px-1 py-2"
-        >
-          {/* Why: Settings deep-links can target this subsection directly from
-          the status-bar account switcher. Keeping a stable DOM anchor here
-          avoids dumping the user at the top of General and making them hunt
-          for the actual Codex account controls. */}
-          <div className="flex items-center justify-between gap-3">
-            <div className="space-y-0.5">
-              <Label>Accounts</Label>
-              <p className="text-xs text-muted-foreground">
-                Add a Codex account to use it in Orca.
-              </p>
-            </div>
-            <Button
-              variant="outline"
-              size="xs"
-              onClick={() =>
-                void runCodexAccountAction('adding', () => window.api.codexAccounts.add())
-              }
-              disabled={codexAction !== 'idle'}
-              className="gap-1.5"
-            >
-              {codexAction === 'adding' ? (
-                <Loader2 className="size-3 animate-spin" />
-              ) : (
-                <Plus className="size-3" />
-              )}
-              Add Account
-            </Button>
-          </div>
-
-          {codexAccounts.accounts.length === 0 ? (
-            <div className="rounded-md border border-dashed border-border/70 px-3 py-4 text-xs text-muted-foreground">
-              No managed Codex accounts yet. Orca will use your system default Codex login until you
-              add one here.
-            </div>
-          ) : (
-            <div className="space-y-2">
-              <button
-                type="button"
-                onClick={() =>
-                  void runCodexAccountAction('select:system', () =>
-                    window.api.codexAccounts.select({ accountId: null })
-                  )
-                }
-                disabled={codexAction !== 'idle'}
-                className={`flex w-full items-center justify-between gap-3 rounded-md border px-3 py-2.5 text-left transition-colors ${
-                  codexAccounts.activeAccountId === null
-                    ? 'border-foreground/20 bg-accent/15'
-                    : 'border-border/70 hover:border-border hover:bg-accent/8'
-                } disabled:cursor-default disabled:opacity-100`}
-              >
-                <div className="flex min-w-0 flex-1 flex-col gap-0.5">
-                  <div className="flex min-w-0 items-center gap-2">
-                    <span className="truncate text-sm font-medium">System default</span>
-                    {codexAccounts.activeAccountId === null ? (
-                      <Badge
-                        variant="outline"
-                        className="h-4 shrink-0 rounded px-1.5 text-[10px] font-medium leading-none text-foreground/80"
-                      >
-                        Active
-                      </Badge>
-                    ) : null}
-                  </div>
-                  <span className="truncate text-[11px] text-muted-foreground">
-                    Use your current system Codex login.
-                  </span>
-                </div>
-              </button>
-              {codexAccounts.accounts.map((account) => {
-                const isActive = codexAccounts.activeAccountId === account.id
-                const isReauthing = codexAction === `reauth:${account.id}`
-                const isRemoving = codexAction === `remove:${account.id}`
-                const isBusy = codexAction !== 'idle'
-
-                return (
-                  <button
-                    key={account.id}
-                    type="button"
-                    onClick={() =>
-                      void runCodexAccountAction(`select:${account.id}`, () =>
-                        window.api.codexAccounts.select({ accountId: account.id })
-                      )
-                    }
-                    disabled={isBusy}
-                    className={`flex w-full items-center justify-between gap-3 rounded-md border px-3 py-2.5 text-left transition-colors ${
-                      isActive
-                        ? 'border-foreground/20 bg-accent/15'
-                        : 'border-border/70 hover:border-border hover:bg-accent/8'
-                    }`}
-                  >
-                    <div className="flex w-full items-center justify-between gap-3 max-md:flex-col max-md:items-start">
-                      <div className="flex min-w-0 flex-1 flex-col gap-0.5">
-                        <div className="flex min-w-0 items-center gap-2">
-                          <span className="truncate text-sm font-medium">{account.email}</span>
-                          {isActive ? (
-                            <Badge
-                              variant="outline"
-                              className="h-4 shrink-0 rounded px-1.5 text-[10px] font-medium leading-none text-foreground/80"
-                            >
-                              Active
-                            </Badge>
-                          ) : null}
-                        </div>
-                        <div className="flex min-w-0 items-center gap-1.5 text-[11px] text-muted-foreground max-sm:flex-wrap">
-                          {account.workspaceLabel ? (
-                            <span className="truncate">{account.workspaceLabel}</span>
-                          ) : null}
-                          {account.workspaceLabel ? (
-                            <span className="shrink-0 opacity-50">•</span>
-                          ) : null}
-                          <span className="shrink-0">
-                            {formatAccountTimestamp(account.lastAuthenticatedAt)}
-                          </span>
-                        </div>
-                      </div>
-
-                      <div className="flex shrink-0 items-center justify-end gap-1 max-md:w-full max-md:flex-wrap">
-                        {/* Why: selecting an account is the primary action in this row.
-                        Keeping maintenance actions visually lighter prevents re-auth/remove
-                        controls from overpowering the selection affordance in a dense list. */}
-                        <Button
-                          variant="ghost"
-                          size="xs"
-                          onClick={(event) => {
-                            event.stopPropagation()
-                            void runCodexAccountAction(`reauth:${account.id}`, () =>
-                              window.api.codexAccounts.reauthenticate({ accountId: account.id })
-                            )
-                          }}
-                          disabled={isBusy}
-                          className="h-6 px-2 text-muted-foreground hover:text-foreground"
-                        >
-                          {isReauthing ? (
-                            <Loader2 className="size-3 animate-spin" />
-                          ) : (
-                            <RefreshCw className="size-3" />
-                          )}
-                          Re-authenticate
-                        </Button>
-                        <Button
-                          variant="ghost"
-                          size="xs"
-                          onClick={(event) => {
-                            event.stopPropagation()
-                            setRemoveAccountId(account.id)
-                          }}
-                          disabled={isBusy}
-                          className="h-6 px-2 text-muted-foreground hover:text-destructive"
-                        >
-                          {isRemoving ? (
-                            <Loader2 className="size-3 animate-spin" />
-                          ) : (
-                            <Trash2 className="size-3" />
-                          )}
-                          Remove
-                        </Button>
-                      </div>
-                    </div>
-                  </button>
-                )
-              })}
-            </div>
-          )}
-        </SearchableSetting>
-      </section>
-    ) : null,
-    matchesSettingsSearch(searchQuery, GENERAL_GEMINI_SEARCH_ENTRIES) ? (
-      <section key="gemini" id="general-gemini" className="space-y-4 scroll-mt-6">
-        <div className="space-y-1">
-          <h3 className="text-sm font-semibold">Gemini</h3>
-          <p className="text-xs text-muted-foreground">Configure Gemini provider settings.</p>
-        </div>
-
-        <SearchableSetting
-          title="Use Gemini CLI credentials"
-          description="Extracts OAuth credentials from your local Gemini CLI installation to authenticate with Google. This uses credentials issued to the Gemini CLI app, not Orca. May break if Google updates the CLI. Use at your own risk."
-          keywords={[
-            'gemini',
-            'cli',
-            'oauth',
-            'credentials',
-            'experimental',
-            'rate limit',
-            'status bar'
-          ]}
-          className="flex items-center justify-between gap-4 px-1 py-2"
-        >
-          <div className="space-y-0.5">
-            <Label>Use Gemini CLI credentials (experimental)</Label>
-            <p className="text-xs text-muted-foreground">
-              Extracts OAuth credentials from your local Gemini CLI installation to authenticate
-              with Google. This uses credentials issued to the Gemini CLI app, not Orca. May break
-              if Google updates the CLI. Use at your own risk.
-            </p>
-          </div>
-          <button
-            role="switch"
-            aria-checked={settings.geminiCliOAuthEnabled}
-            onClick={() =>
-              updateSettings({
-                geminiCliOAuthEnabled: !settings.geminiCliOAuthEnabled
-              })
-            }
-            className={`relative inline-flex h-5 w-9 shrink-0 cursor-pointer items-center rounded-full border border-transparent transition-colors ${
-              settings.geminiCliOAuthEnabled ? 'bg-foreground' : 'bg-muted-foreground/30'
-            }`}
-          >
-            <span
-              className={`pointer-events-none block size-3.5 rounded-full bg-background shadow-sm transition-transform ${
-                settings.geminiCliOAuthEnabled ? 'translate-x-4' : 'translate-x-0.5'
-              }`}
-            />
-          </button>
-        </SearchableSetting>
-      </section>
-    ) : null,
-    matchesSettingsSearch(searchQuery, GENERAL_OPENCODE_SEARCH_ENTRIES) ? (
-      <section key="opencode-go" id="general-opencode-go" className="space-y-4 scroll-mt-6">
-        <div className="space-y-1">
-          <h3 className="text-sm font-semibold">OpenCode Go</h3>
-          <p className="text-xs text-muted-foreground">Configure OpenCode Go provider settings.</p>
-        </div>
-
-        <SearchableSetting
-          title="OpenCode Go Session Cookie"
-          description="Paste your opencode.ai session cookie for rate limit fetching."
-          keywords={['opencode', 'cookie', 'session', 'rate limit', 'status bar']}
-          className="space-y-2"
-        >
-          <Label>OpenCode Go session cookie</Label>
-          <div className="flex gap-2">
-            <Input
-              type="password"
-              value={settings.opencodeSessionCookie}
-              onChange={(e) => updateSettings({ opencodeSessionCookie: e.target.value })}
-              placeholder="Fe26.2**… token or auth=Fe26.2**… header"
-              spellCheck={false}
-              className="flex-1 text-xs"
-            />
-            {settings.opencodeSessionCookie && (
-              <Button
-                variant="ghost"
-                size="xs"
-                onClick={() => updateSettings({ opencodeSessionCookie: '' })}
-                className="h-7 shrink-0 text-xs text-muted-foreground hover:text-foreground"
-              >
-                Clear
-              </Button>
-            )}
-          </div>
-          <p className="text-xs text-muted-foreground">
-            Paste either the raw token value (e.g. <code className="text-xs">Fe26.2**…</code>) or
-            the full cookie header (e.g. <code className="text-xs">auth=Fe26.2**…</code>). Find it
-            in your browser&apos;s DevTools → Network → any opencode.ai request → Cookie header.
-          </p>
-        </SearchableSetting>
-
-        <SearchableSetting
-          title="OpenCode Go Workspace ID"
-          description="Optional workspace ID override if the automatic lookup fails."
-          keywords={['opencode', 'workspace', 'id', 'wrk', 'rate limit', 'status bar']}
-          className="space-y-2"
-        >
-          <Label>Workspace ID override</Label>
-          <div className="flex gap-2">
-            <Input
-              type="text"
-              value={settings.opencodeWorkspaceId}
-              onChange={(e) => updateSettings({ opencodeWorkspaceId: e.target.value })}
-              placeholder="wrk_…  (leave blank for automatic lookup)"
-              spellCheck={false}
-              className="flex-1 text-xs"
-            />
-            {settings.opencodeWorkspaceId && (
-              <Button
-                variant="ghost"
-                size="xs"
-                onClick={() => updateSettings({ opencodeWorkspaceId: '' })}
-                className="h-7 shrink-0 text-xs text-muted-foreground hover:text-foreground"
-              >
-                Clear
-              </Button>
-            )}
-          </div>
-          <p className="text-xs text-muted-foreground">
-            Find this in the URL after logging into opencode.ai (e.g.{' '}
-            <code className="text-xs">opencode.ai/workspace/wrk_…/go</code>).
-          </p>
-        </SearchableSetting>
-      </section>
-    ) : null,
     matchesSettingsSearch(searchQuery, GENERAL_UPDATE_SEARCH_ENTRIES) ? (
       <section key="updates" className="space-y-4">
         <div className="space-y-1">
@@ -1261,74 +577,6 @@ export function GeneralPane({ settings, updateSettings }: GeneralPaneProps): Rea
 
   return (
     <div className="space-y-8">
-      <Dialog
-        open={removeAccountId !== null}
-        onOpenChange={(open) => !open && setRemoveAccountId(null)}
-      >
-        <DialogContent showCloseButton={false}>
-          <DialogHeader>
-            <DialogTitle>Remove Codex Account?</DialogTitle>
-            <DialogDescription>
-              Orca will delete the managed Codex home for this saved account. If it is currently
-              active, Orca falls back to the system default Codex login.
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setRemoveAccountId(null)}>
-              Cancel
-            </Button>
-            <Button
-              variant="destructive"
-              onClick={() => {
-                const accountId = removeAccountId
-                if (!accountId) {
-                  return
-                }
-                setRemoveAccountId(null)
-                void runCodexAccountAction(`remove:${accountId}`, () =>
-                  window.api.codexAccounts.remove({ accountId })
-                )
-              }}
-            >
-              Remove Account
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-      <Dialog
-        open={removeClaudeAccountId !== null}
-        onOpenChange={(open) => !open && setRemoveClaudeAccountId(null)}
-      >
-        <DialogContent showCloseButton={false}>
-          <DialogHeader>
-            <DialogTitle>Remove Claude Account?</DialogTitle>
-            <DialogDescription>
-              Orca will delete the managed Claude auth for this saved account. If it is currently
-              active, Orca falls back to the system default Claude login.
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setRemoveClaudeAccountId(null)}>
-              Cancel
-            </Button>
-            <Button
-              variant="destructive"
-              onClick={() => {
-                const accountId = removeClaudeAccountId
-                if (!accountId) {
-                  return
-                }
-                setRemoveClaudeAccountId(null)
-                void runClaudeAccountAction(`remove:${accountId}`, () =>
-                  window.api.claudeAccounts.remove({ accountId })
-                )
-              }}
-            >
-              Remove Account
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
       {visibleSections.map((section, index) => (
         <div key={index} className="space-y-8">
           {index > 0 ? <Separator /> : null}
