+++ src/renderer/src/components/settings/general-search.ts
@@ -65,49 +65,6 @@ export const GENERAL_CACHE_TIMER_SEARCH_ENTRIES: SettingsSearchEntry[] = [
   }
 ]
 
-export const GENERAL_CLAUDE_ACCOUNTS_SEARCH_ENTRIES: SettingsSearchEntry[] = [
-  {
-    title: 'Claude Accounts',
-    description: 'Manage which Claude account Orca uses while preserving shared chat context.',
-    keywords: ['claude', 'account', 'switch', 'active', 'status bar', 'quota']
-  }
-]
-
-export const GENERAL_CODEX_ACCOUNTS_SEARCH_ENTRIES: SettingsSearchEntry[] = [
-  {
-    title: 'Codex Accounts',
-    description: 'Manage which Codex account Orca uses for live rate limit fetching.',
-    keywords: ['codex', 'account', 'rate limit', 'status bar', 'quota']
-  },
-  {
-    title: 'Active Codex Account',
-    description: 'Choose which saved Codex account powers live quota reads.',
-    keywords: ['codex', 'account', 'switch', 'active', 'status bar']
-  }
-]
-
-export const GENERAL_GEMINI_SEARCH_ENTRIES: SettingsSearchEntry[] = [
-  {
-    title: 'Use Gemini CLI credentials',
-    description:
-      'Extracts OAuth credentials from your local Gemini CLI installation to authenticate with Google.',
-    keywords: ['gemini', 'cli', 'oauth', 'credentials', 'experimental', 'rate limit', 'status bar']
-  }
-]
-
-export const GENERAL_OPENCODE_SEARCH_ENTRIES: SettingsSearchEntry[] = [
-  {
-    title: 'OpenCode Go Session Cookie',
-    description: 'Paste your opencode.ai session cookie for rate limit fetching.',
-    keywords: ['opencode', 'cookie', 'session', 'rate limit', 'status bar']
-  },
-  {
-    title: 'OpenCode Go Workspace ID',
-    description: 'Optional workspace ID override if the automatic lookup fails.',
-    keywords: ['opencode', 'workspace', 'id', 'wrk', 'rate limit', 'status bar']
-  }
-]
-
 export const GENERAL_AGENT_SEARCH_ENTRIES: SettingsSearchEntry[] = [
   {
     title: 'Default Agent',
@@ -139,10 +96,6 @@ export const GENERAL_PANE_SEARCH_ENTRIES: SettingsSearchEntry[] = [
   ...GENERAL_EDITOR_SEARCH_ENTRIES,
   ...GENERAL_CLI_SEARCH_ENTRIES,
   ...GENERAL_CACHE_TIMER_SEARCH_ENTRIES,
-  ...GENERAL_CLAUDE_ACCOUNTS_SEARCH_ENTRIES,
-  ...GENERAL_CODEX_ACCOUNTS_SEARCH_ENTRIES,
-  ...GENERAL_GEMINI_SEARCH_ENTRIES,
-  ...GENERAL_OPENCODE_SEARCH_ENTRIES,
   ...GENERAL_UPDATE_SEARCH_ENTRIES,
   ...GENERAL_SUPPORT_SEARCH_ENTRIES
 ]
