+++ src/renderer/src/store/slices/ui.ts
@@ -103,6 +103,7 @@ export type UISlice = {
       | 'shortcuts'
       | 'repo'
       | 'agents'
+      | 'accounts'
       | 'experimental'
       | 'ssh'
     repoId: string | null
